Clean up, renames, fix tests, comments, console logs cleanups
