+++ app/adapters/homepage.js
@@ -1,9 +1,10 @@
 import PageAdapter from './page';
-import { WAGTAIL_MODEL_TYPE } from '../models/home'
+import { WAGTAIL_MODEL_TYPE } from '../models/homepage'
 
 export const DEFAULT_QUERY_PARAMS = {
   type: WAGTAIL_MODEL_TYPE,
 }
 
 export default PageAdapter.extend({
+
 });
