+++ tests/unit/adapters/homepage-test.js
@@ -1,12 +1,12 @@
 import { module, test } from 'qunit';
 import { setupTest } from 'ember-qunit';
 
-module('Unit | Adapter | content collection', function(hooks) {
+module('Unit | Adapter | homepage', function(hooks) {
   setupTest(hooks);
 
   // Replace this with your real tests.
   test('it exists', function(assert) {
-    let adapter = this.owner.lookup('adapter:content-collection');
+    let adapter = this.owner.lookup('adapter:homepage');
     assert.ok(adapter);
   });
 });
