+++ tests/unit/serializers/homepage-test.js
@@ -1,20 +1,20 @@
 import { module, test } from 'qunit';
 import { setupTest } from 'ember-qunit';
 
-module('Unit | Serializer | content collection', function(hooks) {
+module('Unit | Serializer | homepage', function(hooks) {
   setupTest(hooks);
 
   // Replace this with your real tests.
   test('it exists', function(assert) {
     let store = this.owner.lookup('service:store');
-    let serializer = store.serializerFor('content-collection');
+    let serializer = store.serializerFor('homepage');
 
     assert.ok(serializer);
   });
 
   test('it serializes records', function(assert) {
     let store = this.owner.lookup('service:store');
-    let record = store.createRecord('content-collection', {});
+    let record = store.createRecord('homepage', {});
 
     let serializedRecord = record.serialize();
